requirements.txt
python-dotenv==1.0.0
requests==2.31.0
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
coinbase-advanced-py==1.0.2
google-cloud-aiplatform==1.36.0
schedule==1.2.0
